data_structures/getting_started/day11/binary_tree_level_order_traversal_test.go
package day11

import (
	"reflect"
	"testing"
)

// TestLevelOrder 测试二叉树的层序遍历
func TestLevelOrder(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want [][]int
	}{
		{
			name: "01",
			args: args{
				root: &TreeNode{
					Val: 3,
					Left: &TreeNode{
						Val:   9,
						Left:  nil,
						Right: nil,
					},
					Right: &TreeNode{
						Val: 20,
						Left: &TreeNode{
							Val:   15,
							Left:  nil,
							Right: nil,
						},
						Right: &TreeNode{
							Val:   7,
							Left:  nil,
							Right: nil,
						},
					},
				},
			},
			want: [][]int{
				{3},
				{9, 20},
				{15, 7},
			},
		},
		{
			name: "02",
			args: args{root: nil},
			want: nil,
		},
	}
	for _, tt := range tests {
		t.Run(
			tt.name, func(t *testing.T) {
				if got := levelOrder(tt.args.root); !reflect.DeepEqual(got, tt.want) {
					t.Errorf("levelOrder() = %v, want %v", got, tt.want)
				}
			},
		)
	}
}
